Token sent to client if they are logged in. Can use token to access restricted routes. Tested with Postman. Passport local strategy finished. Passport jwt strategy finished. Consolidated routes into API for client frontend and admin for the eventual CMS.
